pyproject.toml
[tool.poetry]
name = "bankfast"
version = "0.1.0"
description = ""
authors = ["kayigmb <kayigm1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
python-multipart = "^0.0.9"
pydantic-settings = "^2.5.2"
fastapi-limiter = "^0.1.6"
redis = "^5.0.8"
passlib = "^1.7.4"
sqlmodel = "^0.0.22"
sse-starlette = "^2.1.3"
aiortc = "^1.9.0"
websockets = "^13.1"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
